discord_agent/pyproject.toml
[tool.poetry]
name = "discord-agent"
version = "0.1.0"
description = ""
authors = ["interstellarninja <ninjainterstellar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
python-dotenv = "^1.0.1"
discord = "2.3.2"
tiktoken = "0.8.0"
pyyaml = "6.0.2"
sentence-transformers = "3.3.1"
nltk = "3.8.1"
faiss-cpu = "1.9.0.post1"
PyGithub = "2.5.0"
markdown2 = "2.5.0"
Pillow = "11.0.0"
aiohttp = "3.11.6"
requests = "2.32.3"
colorama = "0.4.6"
names = "0.3.0"
openai = "1.56.1"
pydantic = "2.8.2"
tenacity = "9.0.0"
anthropic = "0.40.0"
pydantic-settings = "2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
